common: add CONFIG_PROMPT to Kconfig
This option specifies the string printed on the console to prompt for
user input.  We have use CONFIG_SYS_PROMPT but the prefix CONFIG_SYS_
seems unsuitable for this configuration because users are allowed to
choose their favorate prompt.

This commit attempts to rename CONFIG_SYS_PROMPT to CONFIG_PROMPT
and to move it to Kconfig.  Each board maintainer is expected to add
CONFIG_PROMPT to his defconfig file if he needs to use a non-default
prompt and then remove a CONFIG_SYS_PROMPT define from his header file.

If CONFIG_SYS_PROMPT is defined in the board header file, it takes
precedence over CONFIG_PROMPT defined in Kconfig.

Signed-off-by: Masahiro Yamada <yamada.m@jp.panasonic.com>