tools: get-toolchains: a tool to get cross-tools for all architectures
When we send patches, we are supposed to test them by build utilities
such as MAKEALL, buildman.  When we want to test global changes, the
first hurdle is, I think, to collect toolchains for all the architectures.

We have some documents about build utilities, but I have not seen any
official information about how to get the suitable cross-tools.
Of course, it is possible to build them from sources, but it is not
necessarily feasible.

Fortunately, the kernel.org site provides us pre-built toolchains, but
some architectures are missing.  Also, some boards fail to build with
the kernel.org tools.  We sometimes see, "where can I get the compiler
for this architecture?" things on the ML.  We should be able to prepare
cross-compilers more easily.

It is true that buildman provides --fetch-arch option for downloading
kernel.org toolchains, but it does not have access to others.  And what
we really want to know is most likely how to get compilers for such minor
architectures as kernel.org does not provide.

This tool intends to be more generic design without hard-coding such
kernel.org things.

To achieve that, this tool consists of two files:
Python script (this file) and the database file containing URLs of tarballs.

We just need to update the latter when new version compilers are released
(or better compilers are found.)  The file is in the form of RFC 822 for
easier editing.

The script only uses Python libraries, not relies on external programs
although it displays wget-like log when downloading tarballs.  :-)

This is RFC because I am thinking it can be more brushed up.
If the basis idea is OK, I will improve code, add more comments.

Note this script is written in Python 3 and only works on Python 3.3
or later.  I do not think it is too much limitation, but some popular
distributions under support might include older version.  For example,
looks like Ubuntu 12.04 LTS is shipped with Python 3.2.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Series-cc: sjg
Series-cc: Joe Hershberger <joe.hershberger@ni.com>
Series-cc: trini
Series-prefix: RFC