mtd: nand: denali: allow to override max_banks from DT property
The previous commit ("mtd: nand: denali: max_banks calculation
changed in revision 5.1") supported the new encoding of the "n_banks"
bits of the "features" register, but there is an unfortunate case
that is not covered by that commit.

Panasonic (its System LSI division is now Socionext) bought several
versions of this IP.  The IP released for Panasonic around Feb. 2012
is revision 5 and uses the old encoding for n_banks (2 << n_banks).
While the one released around Nov. 2012 is also revision 5, but it
uses the new encoding (1 << n_banks).

The revision register cannot distinguish these two incompatible
hardware.  I guess this IP series is not well-organized.

This commit works around the problem by allowing to specify the
max_banks with a CONFIG option.  If the option is set to zero
(this is default), the max_banks is auto-detected based on the
information from the hardware registers.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>