usb: fix usb_stor_read/write on DM
Prior to DM, we could not enable different types of USB controllers
at the same time.  DM was supposed to loosen the limitation.  It is
true that we can compile drivers, but they do not work.

For example, if EHCI is enabled, xHCI fails as follows:

  => usb read 82000000 0 2000

  USB read: device 0 block # 0, count 8192 ... WARN halted endpoint, queueing URB anyway.
  Unexpected XHCI event TRB, skipping... (3fb54010 00000001 13000000 01008401)
  BUG: failure at drivers/usb/host/xhci-ring.c:489/abort_td()!
  BUG!
  ### ERROR ### Please RESET the board ###

The cause of the error seems the following code:

  #ifdef CONFIG_USB_EHCI_HCD
  /*
   * The U-Boot EHCI driver can handle any transfer length as long as there is
   * enough free heap space left, but the SCSI READ(10) and WRITE(10) commands are
   * limited to 65535 blocks.
   */
  #define USB_MAX_XFER_BLK	65535
  #else
  #define USB_MAX_XFER_BLK	20
  #endif

To fix the problem, choose the chunk size at run-time for CONFIG_BLK.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Series-version: 2
Series-changes: 2
  - Run-time detection of host if